+++ packages/selector-specificity/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to Selector Specificity
 
+### Unreleased (patch)
+
+- Prevent mutation of selectors with An+B microsyntax (e.g. `:nth-child(2n+1 of .foo)`) during specificity calculation
+
 ### 3.1.0
 
 _May 11, 2024_